sperate netbench function to another file
